requirements.txt
aiohttp
async-timeout
attrs
chardet
httplib2
idna
idna-ssl
multidict
pprint
pyasn1
pyasn1-modules
rsa
six
telepot
uritemplate
yarl
pymongo
maclookup
pillow
bs4
pypdf2
exifread
python-docx